Move colony task tests to own test file
