+++ truffle.js
@@ -33,8 +33,6 @@ module.exports = {
   },
   compilers: {
     solc: {
-      version: "0.4.23",
-      docker: true,
       settings: {
         optimizer: {
           enabled: true,
